+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to **Dangl.Calculator** are documented here.
 
+## v2.2.0:
+- Dropped tests for `net6.0` and addes tests for `net8.0`
+- Added a new overload to the calculator which accepts an options object. Additionally, it is now possible to configure a special case that applies different handling for formulas that start with a negative power, e.g. `-2^2`. This would typically be parsed as `(-2)^2`, but with the new option, it can be configured to be parsed as `-(2^2)`. It is not recommended to use this option, since most other software (e.g. Microsoft Excel and Google Sheets) do also not behave in that way.
+- Square `[]` and curly `{}` parentheses are now also valid for formulas. They are only used to group expressions, and implicit multiplication like `2(3)=6` is not allowed with those
+
 ## v2.1.0:
 - Added support for ranges, e.g. `#START..#END` to implement custom sum logic